+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.2.1 (2022-05-24)
+
+- Added a property to specify the time context to use for the output images indices
+- Added a property to set the start frame on video (videoFirstFrameIndex)
+- Added a property to control the number of digits in the frame numbers (frameDigitsPadding)
+
+- Added an add-on preferences property to toggle the visibility of the main UI panel in the 3D View
+
+### Code
+
+- Add-on Preference property named mediaFirstFrameIsZero has been removed, it is now
+  advantageously replaced by videoFirstFrameIndex
+
 # 1.1.2 (2022-05-11)
 
 - Fix on the get latest release version path